Fix compilation on JDK 8
